Add solution file and .gitignore
A new Microsoft Visual Studio solution file is added to the FSharp.EventSourcing.ToDoAPI project. Also included in this commit is a new .gitignore file with general ignores such as 'bin/', 'obj/', '*.user', '*.suo', etc, to prevent unnecessary tracking of files and directories that are generally the outcome of our personal local development environment.