Add .idea/** to .gitignore
This commit includes the addition of the .idea/** line into the .gitignore file. The purpose of this change is to ensure that all files and folders within the .idea directory, which are usually machine or IDE-specific settings, are ignored by Git to prevent unnecessary and potentially conflicting changes from being tracked.