Refactor task creation and state updates in tests
A new module, Features, has been introduced in the FSharp.EventSourcing.ToDo.Domain which contains functionality for creating tasks and updating state. The tests in CreateTaskTests.fs have been updated to use these new functions, streamlining the code and putting most of the logic in the new Features module. This keeps the tests clean and focused on checking functionality.